Refactor main.c and set guest-CR3 load/store exit control
